move shared-components to app/components
