+++ app/benchmarking/components/benchmarkNewSessionFlow.tsx
@@ -2,14 +2,14 @@
 import { useRouter } from 'next/navigation';
 import React, { useLayoutEffect, useState } from 'react';
 import { CookbooksSelection } from '@/app/benchmarking/components/cookbooksSelection';
+import { EndpointSelectVew } from '@/app/benchmarking/components/endpointsSelector';
 import { CookbooksProvider } from '@/app/benchmarking/contexts/cookbooksContext';
 import { Icon, IconName } from '@/app/components/IconSVG';
+import { MainSectionSurface } from '@/app/components/mainSectionSurface';
+import { Modal } from '@/app/components/modal';
 import SimpleStepsIndicator from '@/app/components/simpleStepsIndicator';
 import { NewEndpointForm } from '@/app/endpoints/(edit)/newEndpointForm';
-import { EndpointSelectVew } from '@/app/benchmarking/components/endpointsSelector';
 import { colors } from '@/app/views/shared-components/customColors';
-import { MainSectionSurface } from '@/app/views/shared-components/mainSectionSurface/mainSectionSurface';
-import { Modal } from '@/app/views/shared-components/modal/modal';
 import {
   addBenchmarkModels,
   removeBenchmarkModels,
