+++ app/benchmarking/components/cookbooksSelection.tsx
@@ -1,10 +1,10 @@
 import dynamic from 'next/dynamic';
 import { useEffect, useState } from 'react';
+import { PopupSurface } from '@/app/components/popupSurface';
 import { calcTotalPromptsAndEstimatedTime } from '@/app/lib/cookbookUtils';
 import { useGetCookbooksQuery } from '@/app/services/cookbook-api-service';
 import { colors } from '@/app/views/shared-components/customColors';
 import { LoadingAnimation } from '@/app/views/shared-components/loadingAnimation';
-import { PopupSurface } from '@/app/views/shared-components/popupSurface/popupSurface';
 import { TabsMenu, TabItem } from '@/app/views/shared-components/tabsMenu';
 import {
   addBenchmarkCookbooks,
