+++ app/benchmarking/recipes/page.tsx
@@ -1,7 +1,7 @@
 import { RecipesViewList } from '@/app/benchmarking/recipes/recipesViewList';
+import { MainSectionSurface } from '@/app/components/mainSectionSurface';
 import { ApiResult, processResponse } from '@/app/lib/http-requests';
 import { colors } from '@/app/views/shared-components/customColors';
-import { MainSectionSurface } from '@/app/views/shared-components/mainSectionSurface/mainSectionSurface';
 import config from '@/moonshot.config';
 export const dynamic = 'force-dynamic';
