+++ app/redteaming/(fullscreen)/components/viewBookmarksModal.tsx
@@ -2,11 +2,11 @@
 import React from 'react';
 import { getAllBookmarks } from '@/actions/getAllBookmarks';
 import { Icon, IconName } from '@/app/components/IconSVG';
+import { Modal } from '@/app/components/modal';
 import { TextInput } from '@/app/components/textInput';
 import { formatDate } from '@/app/lib/date-utils';
 import { colors } from '@/app/views/shared-components/customColors';
 import { LoadingAnimation } from '@/app/views/shared-components/loadingAnimation';
-import { Modal } from '@/app/views/shared-components/modal/modal';
 import { ColorCodedTemplateString } from './color-coded-template';
 
 type ViewBookmarksModalProps = {
@@ -193,13 +193,15 @@ function ViewBookmarksModal(props: ViewBookmarksModalProps) {
       secondaryBtnLabel="Cancel"
       onPrimaryBtnClick={!isPending ? handlePrimaryBtnClick : undefined}
       onSecondaryBtnClick={onCloseIconClick}>
-      {isPending ? <LoadingAnimation /> : (
+      {isPending ? (
+        <LoadingAnimation />
+      ) : (
         <main
           className="flex gap-5 mb-3 w-full"
           style={{
             height: 'calc(100% - 90px)',
           }}>
-          {(!isPending && !bookmarks.length) ? (
+          {!isPending && !bookmarks.length ? (
             <p>No bookmarks found</p>
           ) : (
             <div className="flex gap-4 h-full w-full">
@@ -210,7 +212,8 @@ function ViewBookmarksModal(props: ViewBookmarksModalProps) {
               <div className="flex-1">{detailsSection}</div>
             </div>
           )}
-        </main>)}
+        </main>
+      )}
     </Modal>
   );
 }
