+++ app/redteaming/(fullscreen)/error.tsx
@@ -3,8 +3,8 @@
 import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 import { Icon, IconName } from '@/app/components/IconSVG';
+import { Modal } from '@/app/components/modal';
 import { colors } from '@/app/views/shared-components/customColors';
-import { Modal } from '@/app/views/shared-components/modal/modal';
 
 export default function Error({
   error,
