+++ app/utilities/context-strategies/contextStrategiesList.tsx
@@ -2,8 +2,8 @@
 import { useSearchParams } from 'next/navigation';
 import React, { CSSProperties, useState } from 'react';
 import { Icon, IconName } from '@/app/components/IconSVG';
+import { MainSectionSurface } from '@/app/components/mainSectionSurface';
 import { colors } from '@/app/views/shared-components/customColors';
-import { MainSectionSurface } from '@/app/views/shared-components/mainSectionSurface/mainSectionSurface';
 
 interface CustomStyle extends CSSProperties {
   WebkitLineClamp?: string;
